+++ includes/rest/class-classes-ci-node.php
@@ -159,6 +159,7 @@ class Classes_CI_Node extends Service_CI_Node {
 				[
 					'name'        => 'list',
 					'description' => 'List registered classes (with schemas) and formatters.',
+					'capability'  => 'read',
 					'args'        => [],
 					'handler'     => static fn ( Command_Interpreter_Node $self, array $args, array $envelope = [] ): array => self::cmd_list(),
 				],